Create Match StepDefs finished v1.

+++ src/main/java/cat/udl/eps/softarch/geolearning/domain/MatchResult.java
@@ -14,7 +14,7 @@ public class MatchResult extends UriEntity<Integer>{
     private Integer id;
 
     private int result;
-    private float time;
+    private int time;
 
     @ManyToOne
     private Match match;
